+++ cocos/3d/assets/dom-audio-clip.js
@@ -3,7 +3,7 @@
 // @ts-check
 import { _decorator } from "../../core/data/index";
 const { ccclass, property } = _decorator;
-import { AudioClip, AudioSourceType, PlayingState } from "./CCAudioClip";
+import { AudioClip, AudioSourceType, PlayingState } from "./audio-clip";
 
 @ccclass
 export class DOMAudioClip extends AudioClip {
