+++ cocos/3d/assets/texture.js
@@ -2,7 +2,7 @@
 
 // @ts-check
 import { _decorator } from "../../core/data/index";
-const { ccclass, property } = _decorator;
+const { ccclass } = _decorator;
 import Asset from "../../assets/CCAsset";
 
 @ccclass
